requirements.txt
nicegui
fastapi
uvicorn
python-dotenv